+++ changelog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.0] 2020-05-04
 ### Added 
 - Emprical forecast mode.  Leverages county infection and population data to forecast future infections.
+### Changed
+- Changed how ICU length of stay is represented.
 
 ## [1.7.1] - 2020-04-14
 ### Added